+++ src/exchanges.rs
@@ -1,6 +1,9 @@
 #[cfg(feature = "binance")]
 #[cfg_attr(docsrs, doc(cfg(feature = "binance")))]
 pub mod binance;
+#[cfg(feature = "bitbank")]
+#[cfg_attr(docsrs, doc(cfg(feature = "bitbank")))]
+pub mod bitbank;
 #[cfg(feature = "bitflyer")]
 #[cfg_attr(docsrs, doc(cfg(feature = "bitflyer")))]
 pub mod bitflyer;